Refactoring & security checks added
