+++ imports/UI/modals/resetPassword.js
@@ -31,11 +31,14 @@ Template.resetPassword.events({
         const newPassword = form.get('newPassword');
         const confirmNewPassword = form.get('confirmNewPassword');
 
+        $(event.target).addClass("is-loading");  // Add a loading effect to the button
+
         Meteor.call('checkPasswordsInput', {password: newPassword, confirmPassword: confirmNewPassword}, function(error, result){
             if(error){
                 // There is an error in password fields
                 Session.set('message', {type:"header", headerContent:error.reason, style:"is-danger"} );  // Display an error message
                 $('input#newPassword, input#confirmNewPassword').addClass("is-danger");  // Adding a red border to those fields
+                $(event.target).removeClass("is-loading");  // Remove the loading effect of the button
             } else{
                 // New passwords match all criteria, catching the token provided in reset password email
                 const token = Session.get('resetPasswordToken');
@@ -44,6 +47,7 @@ Template.resetPassword.events({
                     if(error){
                         // There was an error while resetting the password
                         Session.set('message', {type:"header", headerContent:error.reason, style:"is-danger"});  // Display an error message
+                        $(event.target).removeClass("is-loading");  // Remove the loading effect of the button
                     } else{
                         // Password was successfully reset, displaying a success message
                         Session.set('message', {type:"header", headerContent:"Mot de passe réinitialisé avec succès", style:"is-success"});
