+++ server/methods/generic.js
@@ -32,24 +32,30 @@ Meteor.methods({
         check(message, String);
 
         // TODO: length verifications
-        // TODO: email verification
-        const from = Rules.email.contactForm.sender;
-        const to = Rules.email.contactForm.receiver;
-        const emailSubject = "Formulaire de contact";
 
-        // Creating the body content of the email
-        const html = `<h3>Sujet: `+ subject +`</h3>
-                      <h4>Adresse email : `+ email +`</h4>
-                      <p>Message : `+ message +`</p>`;
+        Meteor.call('checkEmailInput', {email: email}, function(error, result){
+            if(error){
+                // Email is invalid, throwing an error
+                throw new Meteor.Error('invalidEmailAddress', "Adresse email invalide.");
+            } else{
+                // Email address is valid
+                const from = Rules.email.contactForm.sender;
+                const to = Rules.email.contactForm.receiver;
+                const emailSubject = "Formulaire de contact";
 
-        // Sending email using SendGrid (https://app.sendgrid.com/guide/integrate/langs/nodejs):
+                // Creating the body content of the email
+                const html = `<h3>Sujet: `+ subject +`</h3>
+                              <h4>Adresse email : `+ email +`</h4>
+                              <p>Message : `+ message +`</p>`;
 
-        // Using Twilio SendGrid's v3 Node.js Library (https://github.com/sendgrid/sendgrid-nodejs)
-        const sendGrid = require('@sendgrid/mail');
-        sendGrid.setApiKey(process.env.SENDGRID_CONTACT_API_KEY);
-        sendGrid.send({to: to, from: from, subject: emailSubject, html: html});
+                // Sending email using SendGrid (https://app.sendgrid.com/guide/integrate/langs/nodejs):
 
-        return true;
+                // Using Twilio SendGrid's v3 Node.js Library (https://github.com/sendgrid/sendgrid-nodejs)
+                const sendGrid = require('@sendgrid/mail');
+                sendGrid.setApiKey(process.env.SENDGRID_CONTACT_API_KEY);
+                sendGrid.send({to: to, from: from, subject: emailSubject, html: html});
+            }
+        });
     },
     'setAccountsSettings'(){
